Implement first recursive version of Morton LBVH
This commit introduces the building and traversing of a BVH by using the Morton
codes of primitives. At the current time there exist certain round-off errors.
Therefore we have problems to render too small or too big scenes.
The algorithm itself and the structuring of the whole ray tracer is still not
finished.

See also: #12