Fix function 'load_stl' to recompute zero-normals
In the binary STL file format the normals of every triangle are already given.
But there are some rare cases where the coordinates of these normals are set to
a value near zero. Shading a primitive with a 'zero-normal' results in a black
screen. Therefore we now check the length of the given normal and recompute the
whole normal if this length it too small.
Additionally we need a normalized ray direction for the shading process. Hence,
we changed the function 'primary_ray' to normalize the direction of the
resulting ray.