Reorganize folder and library structure
The main part of the project is now the libray 'spray' which at this time
consists of the two sub-libraries 'chrono' and 'ray_tracer'. They are using
their own directories. The main application of this project is now called
'spray_viewer'. Everything can be configured through a file 'CMakeLists.txt' in
the respective sub-directories.

See also: #8